+++ cli/internal/specs/v0/destination.go
@@ -27,10 +27,10 @@ type Destination struct {
 func (d *Destination) GetWarnings() Warnings {
 	warnings := make(map[string]string)
 	if d.BatchSize != 0 {
-		warnings["batch_size"] = "batch_size in the top-level spec is deprecated and will be removed in a future release. Please use the plugin-spec batch_size option instead."
+		warnings["batch_size"] = "batch_size in the top-level spec is deprecated and will be removed in a future release. Please use the plugin-spec `batch_size` option instead."
 	}
 	if d.BatchSizeBytes != 0 {
-		warnings["batch_size_bytes"] = "batch_size_bytes in the top-level spec is deprecated and will be removed in a future release. Please use the plugin-spec batch_size_bytes option instead."
+		warnings["batch_size_bytes"] = "batch_size_bytes in the top-level spec is deprecated and will be removed in a future release. Please use the plugin-spec `batch_size_bytes` option instead."
 	}
 	return warnings
 }