+++ plugins/destination/gremlin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [2.2.4](https://github.com/cloudquery/cloudquery/compare/plugins-destination-gremlin-v2.2.3...plugins-destination-gremlin-v2.2.4) (2023-08-08)
+
+
+### Bug Fixes
+
+* **deps:** Update AWS modules ([#12584](https://github.com/cloudquery/cloudquery/issues/12584)) ([bc4d46d](https://github.com/cloudquery/cloudquery/commit/bc4d46d5edae7ac10899a0580d4d7bdf14d443be))
+* **deps:** Update AWS modules ([#12772](https://github.com/cloudquery/cloudquery/issues/12772)) ([a0abac5](https://github.com/cloudquery/cloudquery/commit/a0abac5811381318957aefd1228b4a91610c71f1))
+* **deps:** Update github.com/apache/arrow/go/v13 digest to 112f949 ([#12659](https://github.com/cloudquery/cloudquery/issues/12659)) ([48d73a9](https://github.com/cloudquery/cloudquery/commit/48d73a93e678994f43171c363f5a75c29547b0b9))
+* **deps:** Update github.com/cloudquery/arrow/go/v13 digest to 3452eb0 ([#12595](https://github.com/cloudquery/cloudquery/issues/12595)) ([c1c0949](https://github.com/cloudquery/cloudquery/commit/c1c09490b17f2e64435e05d745890cdb8b22310d))
+* **deps:** Update github.com/cloudquery/arrow/go/v13 digest to f53878d ([#12778](https://github.com/cloudquery/cloudquery/issues/12778)) ([6f5d58e](https://github.com/cloudquery/cloudquery/commit/6f5d58e3b84d3c76b1d1a3d6c5a488f77995a057))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.2.4 ([#12718](https://github.com/cloudquery/cloudquery/issues/12718)) ([f059a15](https://github.com/cloudquery/cloudquery/commit/f059a159a2ee406ab2b0a33792c244cd217025a6))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.2.5 ([#12731](https://github.com/cloudquery/cloudquery/issues/12731)) ([d267239](https://github.com/cloudquery/cloudquery/commit/d267239aa3aca5f94bd36a8db1ec0d9f7dc0865f))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.2.6 ([#12799](https://github.com/cloudquery/cloudquery/issues/12799)) ([fb0e0d7](https://github.com/cloudquery/cloudquery/commit/fb0e0d75ab010f421c834e58d93676de76fcb423))
+
 ## [2.2.3](https://github.com/cloudquery/cloudquery/compare/plugins-destination-gremlin-v2.2.2...plugins-destination-gremlin-v2.2.3) (2023-07-25)